refactor: split oversized modules and cut function complexity (MOD-001) (#743)
* refactor: split oversized modules and cut function complexity (MOD-001)

The dev->main promotion (PR #740) failed SonarCloud's aces-strict quality gate
with 9 new violations: 5 files over the 500-line S104 cap and 4 functions over
cyclomatic complexity 10. Split each oversized module along a cohesive seam and
decomposed each over-complex function into helpers. Moved symbols are re-imported
into their original modules so every public import path and __all__ stays stable.
No behaviour change; full nox verify green.

Splits:
- aces_sdl.parser -> _model_diagnostics (pydantic-error diagnostic rendering)
- aces_sdl._runtime_service_families -> _runtime_service_family_registry (registry data)
- aces_sdl._yaml_loader -> _mapping_key_analyzer (mapping-key validation walker)
- aces_operations._evidence_run_validation -> _evidence_run_realization
- aces_operations.libvirt_evidence_run -> _evidence_run_native (EvidenceCheck /
  LibvirtEvidenceRunConfig relocated to _evidence_run_types to break the cycle)
- aces_backend_libvirt.techvault_concerns -> techvault_plan_admission
- aces_backend_libvirt.techvault_native -> _techvault_native_ops

Complexity reductions (all now <=10):
- _evidence_run_realization._validate_unrealized_substrate (11)
- _mapping_key_analyzer._walk_mapping_entry (13)
- techvault_native.TechVaultNativeLibvirtDriver.__post_init__ (12)
- provisioner.LibvirtProvisioner._drive (14)

* refactor: make LibvirtProvisioner._active_addresses static (S2325)

The extracted _active_addresses helper does not use instance state; SonarCloud
flagged python:S2325. Make it a staticmethod (called via the instance in _drive).
