+++ implementations/python/packages/aces_backend_libvirt/techvault_native.py
@@ -8,17 +8,30 @@ substrate boundary here is libvirt.
 
 from __future__ import annotations
 
-from collections.abc import Callable, Mapping, Sequence
+from collections.abc import Mapping, Sequence
 from contextlib import suppress
 from dataclasses import dataclass, field
 from pathlib import Path
-from typing import TYPE_CHECKING, ClassVar, Protocol, cast
+from typing import TYPE_CHECKING, ClassVar, cast
 from urllib.parse import urlsplit
 
-from aces_contracts.diagnostics import Diagnostic, Severity
+from aces_contracts.diagnostics import Diagnostic
 
 from ._techvault_native_helpers import default_connector as _default_connector
 from ._techvault_native_helpers import default_kernel_path as _default_kernel_path
+from ._techvault_native_ops import (
+    _CODE_OPERATION_FAILED,
+    _CODE_OWNERSHIP_CONFLICT,
+    _CODE_READBACK_FAILED,
+    _CODE_RESIDUAL_STATE,
+    _CODE_UNAVAILABLE,
+    _DEFAULT_CONNECTION_URI,
+    _artifact_token,
+    _call,
+    _diagnostic,
+    _ensure_name_available,
+    _NativeResource,
+)
 
 if TYPE_CHECKING:
     from aces_contracts.realization_envelope import BackendRealizationEnvelopeModel
@@ -31,7 +44,7 @@ from .driver import (
     NetworkSpec,
     RealizationObservation,
 )
-from .drivers.libvirt import Connector, _aces_uuid, _error_code, _existing_uuid
+from .drivers.libvirt import Connector, _aces_uuid, _existing_uuid
 from .envelopes import load_libvirt_realization_envelope
 from .techvault_appliance import (
     BusyboxInitramfsBuilder,
@@ -82,13 +95,6 @@ from .techvault_probe import (
     native_soc_readback,
 )
 
-_DOMAIN = "runtime"
-_CODE_OPERATION_FAILED = "libvirt-backend.techvault-native.operation-failed"
-_CODE_OWNERSHIP_CONFLICT = "libvirt-backend.techvault-native.ownership-conflict"
-_CODE_READBACK_FAILED = "libvirt-backend.techvault-native.readback-failed"
-_CODE_RESIDUAL_STATE = "libvirt-backend.techvault-native.residual-state"
-_CODE_UNAVAILABLE = "libvirt-backend.techvault-native.unavailable"
-_DEFAULT_CONNECTION_URI = "qemu:///system"
 __all__ = [
     "BusyboxInitramfsBuilder",
     "NativeLibvirtProbe",
@@ -98,14 +104,6 @@ __all__ = [
     "expected_surface",
     "native_soc_readback",
 ]
-
-
-class _NativeResource(Protocol):
-    def create(self) -> None: ...
-
-    def destroy(self) -> None: ...
-
-    def undefine(self) -> None: ...
 
 
 @dataclass
@@ -126,11 +124,23 @@ class TechVaultNativeLibvirtDriver:
     last_snapshot: dict[str, object] = field(default_factory=dict)
 
     def __post_init__(self) -> None:
+        self._validate_connection_uri()
+        self._validate_appliance_flags()
+        self.state_dir = Path(self.state_dir)
+        self.kernel_path = Path(self.kernel_path) if self.kernel_path is not None else _default_kernel_path()
+        self.connector = self.connector or _default_connector
+        self._names: dict[str, str] = {}
+        self._realized: set[str] = set()
+        self._artifacts: dict[str, tuple[Path, ...]] = {}
+
+    def _validate_connection_uri(self) -> None:
         if not self.connection_uri or not self.connection_uri.strip():
             raise ValueError("TechVaultNativeLibvirtDriver connection_uri must be non-empty.")
         parsed_uri = urlsplit(self.connection_uri)
         if parsed_uri.username is not None or parsed_uri.password is not None:
             raise ValueError("TechVaultNativeLibvirtDriver connection URI must not carry credentials.")
+
+    def _validate_appliance_flags(self) -> None:
         if not self.name_prefix or not self.name_prefix.strip():
             raise ValueError("TechVaultNativeLibvirtDriver name_prefix must be non-empty.")
         if self.define_only:
@@ -140,12 +150,6 @@ class TechVaultNativeLibvirtDriver:
         safe_prefix = _safe_name(self.name_prefix, fallback="aces-techvault", prefix="")
         if safe_prefix != self.name_prefix:
             raise ValueError("TechVaultNativeLibvirtDriver name_prefix must already be libvirt-safe.")
-        self.state_dir = Path(self.state_dir)
-        self.kernel_path = Path(self.kernel_path) if self.kernel_path is not None else _default_kernel_path()
-        self.connector = self.connector or _default_connector
-        self._names: dict[str, str] = {}
-        self._realized: set[str] = set()
-        self._artifacts: dict[str, tuple[Path, ...]] = {}
 
     def realize(
         self,
@@ -557,43 +561,3 @@ class TechVaultNativeLibvirtDriver:
             return self._destroy_one(connection, lookup_method, address)
         except Exception:
             return False
-
-
-def _call(connection: object, method_name: str, payload: str) -> _NativeResource:
-    method = cast(Callable[[str], _NativeResource], getattr(connection, method_name))
-    return method(payload)
-
-
-def _ensure_name_available(connection: object, method_name: str, name: str, address: str) -> None:
-    method = getattr(connection, method_name, None)
-    if not callable(method):
-        raise RuntimeError("native lookup is unavailable")
-    try:
-        native = method(name)
-    except KeyError:
-        return
-    except Exception as exc:
-        if _error_code(exc) in {42, 43}:
-            return
-        raise
-    if _existing_uuid(native) != _aces_uuid(address):
-        raise _OwnershipConflict(address)
-    raise RuntimeError("owned native object already exists for CREATE")
-
-
-def _artifact_token(address: str) -> str:
-    return _aces_uuid(address).replace("-", "")
-
-
-_MESSAGES = {
-    _CODE_UNAVAILABLE: "Libvirt connection is unavailable for native TechVault realization.",
-    _CODE_RESIDUAL_STATE: "TechVault rollback could not verify cleanup for '{address}'; residual state may remain.",
-    _CODE_OWNERSHIP_CONFLICT: "Native object for '{address}' is not owned by that ACES address; refusing mutation.",
-    _CODE_READBACK_FAILED: "Native libvirt TechVault readback for '{address}' did not succeed.",
-}
-_DEFAULT_MESSAGE = "Native libvirt TechVault operation for '{address}' did not succeed."
-
-
-def _diagnostic(code: str, address: str) -> Diagnostic:
-    message = _MESSAGES.get(code, _DEFAULT_MESSAGE).format(address=address)
-    return Diagnostic(code=code, domain=_DOMAIN, address=address, message=message, severity=Severity.ERROR)
